+++ Haxl/Core/Monad.hs
@@ -41,6 +41,14 @@ module Haxl.Core.Monad
     GenHaxl(..)
   , Result(..)
 
+    -- * Writes (for debugging only)
+  , WriteTree(..)
+  , tellWrite
+  , write
+  , flattenWT
+  , appendWTs
+  , mbModifyWLRef
+
     -- * Cont
   , Cont(..)
   , toHaxl
@@ -51,6 +59,7 @@ module Haxl.Core.Monad
   , newIVar
   , newFullIVar
   , getIVar
+  , getIVarWithWrites
   , putIVar
 
     -- * ResultVal
@@ -138,11 +147,11 @@ trace_ _ = id
 -- The environment
 
 -- | The data we carry around in the Haxl monad.
-data Env u = Env
-  { cacheRef     :: {-# UNPACK #-} !(IORef (DataCache (IVar u)))
+data Env u w = Env
+  { cacheRef     :: {-# UNPACK #-} !(IORef (DataCache (IVar u w)))
       -- ^ cached data fetches
 
-  , memoRef      :: {-# UNPACK #-} !(IORef (DataCache (IVar u)))
+  , memoRef      :: {-# UNPACK #-} !(IORef (DataCache (IVar u w)))
       -- ^ memoized computations
 
   , flags        :: !Flags
@@ -169,13 +178,13 @@ data Env u = Env
        -- ^ The set of requests that we have not submitted to data sources yet.
        -- Owned by the scheduler.
 
-  , runQueueRef :: {-# UNPACK #-} !(IORef (JobList u))
+  , runQueueRef :: {-# UNPACK #-} !(IORef (JobList u w))
        -- ^ runnable computations. Things get added to here when we wake up
        -- a computation that was waiting for something.  When the list is
        -- empty, either we're finished, or we're waiting for some data fetch
        -- to return.
 
-  , completions :: {-# UNPACK #-} !(TVar [CompleteReq u])
+  , completions :: {-# UNPACK #-} !(TVar [CompleteReq u w])
        -- ^ Requests that have completed.  Modified by data sources
        -- (via putResult) and the scheduler.  Waiting for this list to
        -- become non-empty is how the scheduler blocks waiting for
@@ -187,21 +196,27 @@ data Env u = Env
        -- some data fetch.
 
   , speculative :: {-# UNPACK #-} !Int
+
+  , writeLogsRef :: {-# UNPACK #-} !(IORef (WriteTree w))
+       -- ^ A log of all writes done as part of this haxl computation. Any
+       -- haxl computation that needs to be memoized runs in its own
+       -- environment so
   }
 
-type Caches u = (IORef (DataCache (IVar u)), IORef (DataCache (IVar u)))
+type Caches u w = (IORef (DataCache (IVar u w)), IORef (DataCache (IVar u w)))
 
-caches :: Env u -> Caches u
+caches :: Env u w -> Caches u w
 caches env = (cacheRef env, memoRef env)
 
 -- | Initialize an environment with a 'StateStore', an input map, a
 -- preexisting 'DataCache', and a seed for the random number generator.
-initEnvWithData :: StateStore -> u -> Caches u -> IO (Env u)
+initEnvWithData :: StateStore -> u -> Caches u w -> IO (Env u w)
 initEnvWithData states e (cref, mref) = do
   sref <- newIORef emptyStats
   pref <- newIORef emptyProfile
   rs <- newIORef noRequests          -- RequestStore
   rq <- newIORef JobNil
+  wl <- newIORef NilWrites
   comps <- newTVarIO []              -- completion queue
   return Env
     { cacheRef = cref
@@ -217,28 +232,64 @@ initEnvWithData states e (cref, mref) = do
     , completions = comps
     , pendingWaits = []
     , speculative = 0
+    , writeLogsRef = wl
     }
 
 -- | Initializes an environment with 'StateStore' and an input map.
-initEnv :: StateStore -> u -> IO (Env u)
+initEnv :: StateStore -> u -> IO (Env u w)
 initEnv states e = do
   cref <- newIORef emptyDataCache
   mref <- newIORef emptyDataCache
   initEnvWithData states e (cref,mref)
 
 -- | A new, empty environment.
-emptyEnv :: u -> IO (Env u)
+emptyEnv :: u -> IO (Env u w)
 emptyEnv = initEnv stateEmpty
 
-speculate :: Env u -> Env u
+speculate :: Env u w -> Env u w
 speculate env@Env{..}
   | speculative == 0 = env { speculative = 1 }
   | otherwise = env
 
-imperative :: Env u -> Env u
+imperative :: Env u w -> Env u w
 imperative env@Env{..}
   | speculative == 1 = env { speculative = 0 }
   | otherwise = env
+
+-- -----------------------------------------------------------------------------
+-- WriteTree
+
+-- | A tree of writes done during a Haxl computation. We could use a simple
+-- list, but this allows us to avoid multiple mappends when concatenating
+-- writes from two haxl computations.
+--
+-- Users should try to treat this data type as opaque, and prefer
+-- to use @flattenWT@ to get a simple list of writes from a @WriteTree@.
+data WriteTree w
+  = NilWrites
+  | SomeWrite w
+  | MergeWrites (WriteTree w) (WriteTree w)
+  deriving (Show)
+
+appendWTs :: WriteTree w -> WriteTree w -> WriteTree w
+appendWTs NilWrites w = w
+appendWTs w NilWrites = w
+appendWTs w1 w2 = MergeWrites w1 w2
+
+-- This function must be called at the end of the Haxl computation to get
+-- a list of writes.
+flattenWT :: WriteTree w -> [w]
+flattenWT = go []
+  where
+    go !ws NilWrites = ws
+    go !ws (SomeWrite w) = w : ws
+    go !ws (MergeWrites w1 w2) = go (go ws w2) w1
+
+-- This is a convenience wrapper over modifyIORef, which only modifies
+-- writeLogsRef IORef, for non NilWrites.
+mbModifyWLRef :: WriteTree w -> IORef (WriteTree w) -> IO ()
+mbModifyWLRef NilWrites _ = return ()
+mbModifyWLRef !wt ref = modifyIORef' ref (`appendWTs` wt)
 
 -- -----------------------------------------------------------------------------
 -- | The Haxl monad, which does several things:
@@ -247,6 +298,11 @@ imperative env@Env{..}
 --    of the scheduler, including unfetched requests and the run queue
 --    of computations.
 --
+-- * It is a writer monad for 'WriteTree'. We strongly advise these be
+--   used only for logs used for debugging. These are not memoized.
+--   Other relevant writes should be returned as function output,
+--   which is the more "functional" way.
+
 --  * It is a concurrency, or resumption, monad. A computation may run
 --    partially and return 'Blocked', in which case the framework should
 --    perform the outstanding requests in the 'RequestStore', and then
@@ -258,11 +314,18 @@ imperative env@Env{..}
 --
 --  * It contains IO, so that we can perform real data fetching.
 --
-newtype GenHaxl u a = GenHaxl
-  { unHaxl :: Env u -> IO (Result u a) }
+newtype GenHaxl u w a = GenHaxl
+  { unHaxl :: Env u w -> IO (Result u w a) }
 
+tellWrite :: w -> GenHaxl u w ()
+tellWrite = write . SomeWrite
 
-instance IsString a => IsString (GenHaxl u a) where
+write :: WriteTree w -> GenHaxl u w ()
+write wt = GenHaxl $ \Env{..} -> do
+  mbModifyWLRef wt writeLogsRef
+  return $ Done ()
+
+instance IsString a => IsString (GenHaxl u w a) where
   fromString s = return (fromString s)
 
 -- -----------------------------------------------------------------------------
@@ -274,13 +337,13 @@ instance IsString a => IsString (GenHaxl u a) where
 -- This could be an ordinary list, but the optimised representation
 -- saves space and time.
 --
-data JobList u
+data JobList u w
  = JobNil
  | forall a . JobCons
-     (Env u)          -- See Note [make withEnv work] below.
-     (GenHaxl u a)
-     {-# UNPACK #-} !(IVar u a)
-     (JobList u)
+     (Env u w)          -- See Note [make withEnv work] below.
+     (GenHaxl u w a)
+     {-# UNPACK #-} !(IVar u w a)
+     (JobList u w)
 
 -- Note [make withEnv work]
 --
@@ -292,12 +355,12 @@ data JobList u
 -- restart it with the correct Env.  So we stash the Env along with
 -- the continuation in the JobList.
 
-appendJobList :: JobList u -> JobList u -> JobList u
+appendJobList :: JobList u w -> JobList u w -> JobList u w
 appendJobList JobNil c = c
 appendJobList c JobNil = c
 appendJobList (JobCons a b c d) e = JobCons a b c $! appendJobList d e
 
-lengthJobList :: JobList u -> Int
+lengthJobList :: JobList u w -> Int
 lengthJobList JobNil = 0
 lengthJobList (JobCons _ _ _ j) = 1 + lengthJobList j
 
@@ -307,43 +370,58 @@ lengthJobList (JobCons _ _ _ j) = 1 + lengthJobList j
 
 -- | A synchronisation point.  It either contains a value, or a list
 -- of computations waiting for the value.
-newtype IVar u a = IVar (IORef (IVarContents u a))
+newtype IVar u w a = IVar (IORef (IVarContents u w a))
 
-data IVarContents u a
-  = IVarFull (ResultVal a)
-  | IVarEmpty (JobList u)
-    -- morally this is a list of @a -> GenHaxl u ()@, but instead of
+data IVarContents u w a
+  = IVarFull (ResultVal a w)
+  | IVarEmpty (JobList u w)
+    -- morally this is a list of @a -> GenHaxl u w ()@, but instead of
     -- using a function, each computation begins with `getIVar` to grab
     -- the value it is waiting for.  This is less type safe but a little
     -- faster (benchmarked with tests/MonadBench.hs).
 
-newIVar :: IO (IVar u a)
+newIVar :: IO (IVar u w a)
 newIVar = IVar <$> newIORef (IVarEmpty JobNil)
 
-newFullIVar :: ResultVal a -> IO (IVar u a)
+newFullIVar :: ResultVal a w -> IO (IVar u w a)
 newFullIVar r = IVar <$> newIORef (IVarFull r)
 
-getIVar :: IVar u a -> GenHaxl u a
-getIVar (IVar !ref) = GenHaxl $ \_env -> do
+getIVar :: IVar u w a -> GenHaxl u w a
+getIVar (IVar !ref) = GenHaxl $ \Env{..} -> do
   e <- readIORef ref
   case e of
-    IVarFull (Ok a) -> return (Done a)
-    IVarFull (ThrowHaxl e) -> return (Throw e)
+    IVarFull (Ok a _wt) -> return (Done a)
+    IVarFull (ThrowHaxl e _wt) -> return (Throw e)
     IVarFull (ThrowIO e) -> throwIO e
     IVarEmpty _ -> return (Blocked (IVar ref) (Cont (getIVar (IVar ref))))
 
 -- Just a specialised version of getIVar, for efficiency in <*>
-getIVarApply :: IVar u (a -> b) -> a -> GenHaxl u b
-getIVarApply (IVar !ref) a = GenHaxl $ \_env -> do
+getIVarApply :: IVar u w (a -> b) -> a -> GenHaxl u w b
+getIVarApply (IVar !ref) a = GenHaxl $ \Env{..} -> do
   e <- readIORef ref
   case e of
-    IVarFull (Ok f) -> return (Done (f a))
-    IVarFull (ThrowHaxl e) -> return (Throw e)
+    IVarFull (Ok f _wt) -> return (Done (f a))
+    IVarFull (ThrowHaxl e _wt) -> return (Throw e)
     IVarFull (ThrowIO e) -> throwIO e
     IVarEmpty _ ->
       return (Blocked (IVar ref) (Cont (getIVarApply (IVar ref) a)))
 
-putIVar :: IVar u a -> ResultVal a -> Env u -> IO ()
+-- Another specialised version of getIVar, for efficiency in cachedComputation
+getIVarWithWrites :: IVar u w a -> GenHaxl u w a
+getIVarWithWrites (IVar !ref) = GenHaxl $ \Env{..} -> do
+  e <- readIORef ref
+  case e of
+    IVarFull (Ok a wt) -> do
+      mbModifyWLRef wt writeLogsRef
+      return (Done a)
+    IVarFull (ThrowHaxl e wt) -> do
+      mbModifyWLRef wt writeLogsRef
+      return  (Throw e)
+    IVarFull (ThrowIO e) -> throwIO e
+    IVarEmpty _ ->
+      return (Blocked (IVar ref) (Cont (getIVarWithWrites (IVar ref))))
+
+putIVar :: IVar u w a -> ResultVal a w -> Env u w -> IO ()
 putIVar (IVar ref) a Env{..} = do
   e <- readIORef ref
   case e of
@@ -353,7 +431,7 @@ putIVar (IVar ref) a Env{..} = do
     IVarFull{} -> error "putIVar: multiple put"
 
 {-# INLINE addJob #-}
-addJob :: Env u -> GenHaxl u b -> IVar u b -> IVar u a -> IO ()
+addJob :: Env u w -> GenHaxl u w b -> IVar u w b -> IVar u w a -> IO ()
 addJob env !haxl !resultIVar (IVar !ref) =
   modifyIORef' ref $ \contents ->
     case contents of
@@ -371,25 +449,26 @@ addJobPanic = error "addJob: not empty"
 -- thrown in the IO monad from exceptions thrown in the Haxl monad, so
 -- that when the result is fetched using getIVar, we can throw the
 -- exception in the right way.
-data ResultVal a
-  = Ok a
-  | ThrowHaxl SomeException
+data ResultVal a w
+  = Ok a (WriteTree w)
+  | ThrowHaxl SomeException (WriteTree w)
   | ThrowIO SomeException
+    -- we get no write logs when an IO exception occurs
 
-done :: ResultVal a -> IO (Result u a)
-done (Ok a) = return (Done a)
-done (ThrowHaxl e) = return (Throw e)
+done :: ResultVal a w -> IO (Result u w a)
+done (Ok a _) = return (Done a)
+done (ThrowHaxl e _) = return (Throw e)
 done (ThrowIO e) = throwIO e
 
-eitherToResultThrowIO :: Either SomeException a -> ResultVal a
-eitherToResultThrowIO (Right a) = Ok a
+eitherToResultThrowIO :: Either SomeException a -> ResultVal a w
+eitherToResultThrowIO (Right a) = Ok a NilWrites
 eitherToResultThrowIO (Left e)
-  | Just HaxlException{} <- fromException e = ThrowHaxl e
+  | Just HaxlException{} <- fromException e = ThrowHaxl e NilWrites
   | otherwise = ThrowIO e
 
-eitherToResult :: Either SomeException a -> ResultVal a
-eitherToResult (Right a) = Ok a
-eitherToResult (Left e) = ThrowHaxl e
+eitherToResult :: Either SomeException a -> ResultVal a w
+eitherToResult (Right a) = Ok a NilWrites
+eitherToResult (Left e) = ThrowHaxl e NilWrites
 
 
 -- -----------------------------------------------------------------------------
@@ -400,10 +479,10 @@ eitherToResult (Left e) = ThrowHaxl e
 -- data source is just to add these to a queue ('completions') using
 -- 'putResult'; the scheduler collects them from the queue and unblocks
 -- the relevant computations.
-data CompleteReq u
+data CompleteReq u w
   = forall a . CompleteReq
       (Either SomeException a)
-      !(IVar u a)  -- IVar because the result is cached
+      !(IVar u w a)  -- IVar because the result is cached
       {-# UNPACK #-} !Int64 -- see Note [tracking allocation in child threads]
 
 
@@ -452,19 +531,19 @@ allocation limit itself, and changing the counter would mess it up.
 -- | The result of a computation is either 'Done' with a value, 'Throw'
 -- with an exception, or 'Blocked' on the result of a data fetch with
 -- a continuation.
-data Result u a
+data Result u w a
   = Done a
   | Throw SomeException
   | forall b . Blocked
-      {-# UNPACK #-} !(IVar u b)
-      (Cont u a)
+      {-# UNPACK #-} !(IVar u w b)
+      (Cont u w a)
          -- ^ The 'IVar' is what we are blocked on; 'Cont' is the
          -- continuation.  This might be wrapped further if we're
          -- nested inside multiple '>>=', before finally being added
-         -- to the 'IVar'.  Morally @b -> GenHaxl u a@, but see
+         -- to the 'IVar'.  Morally @b -> GenHaxl u w a@, but see
          -- 'IVar',
 
-instance (Show a) => Show (Result u a) where
+instance (Show a) => Show (Result u w a) where
   show (Done a) = printf "Done(%s)" $ show a
   show (Throw e) = printf "Throw(%s)" $ show e
   show Blocked{} = "Blocked"
@@ -518,22 +597,22 @@ turn it into a Haxl exception:
 -- O(n^2) complexity for some pathalogical cases - see the "seql" benchmark
 -- in tests/MonadBench.hs.
 -- See "A Smart View on Datatypes", Jaskelioff/Rivas, ICFP'15
-data Cont u a
-  = Cont (GenHaxl u a)
-  | forall b. Cont u b :>>= (b -> GenHaxl u a)
-  | forall b. (b -> a) :<$> (Cont u b)
+data Cont u w a
+  = Cont (GenHaxl u w a)
+  | forall b. Cont u w b :>>= (b -> GenHaxl u w a)
+  | forall b. (b -> a) :<$> (Cont u w b)
 
-toHaxl :: Cont u a -> GenHaxl u a
+toHaxl :: Cont u w a -> GenHaxl u w a
 toHaxl (Cont haxl) = haxl
 toHaxl (m :>>= k) = toHaxlBind m k
 toHaxl (f :<$> x) = toHaxlFmap f x
 
-toHaxlBind :: Cont u b -> (b -> GenHaxl u a) -> GenHaxl u a
+toHaxlBind :: Cont u w b -> (b -> GenHaxl u w a) -> GenHaxl u w a
 toHaxlBind (m :>>= k) k2 = toHaxlBind m (k >=> k2)
 toHaxlBind (Cont haxl) k = haxl >>= k
 toHaxlBind (f :<$> x) k = toHaxlBind x (k . f)
 
-toHaxlFmap :: (a -> b) -> Cont u a -> GenHaxl u b
+toHaxlFmap :: (a -> b) -> Cont u w a -> GenHaxl u w b
 toHaxlFmap f (m :>>= k) = toHaxlBind m (k >=> return . f)
 toHaxlFmap f (Cont haxl) = f <$> haxl
 toHaxlFmap f (g :<$> x) = toHaxlFmap (f . g) x
@@ -542,7 +621,7 @@ toHaxlFmap f (g :<$> x) = toHaxlFmap (f . g) x
 -- -----------------------------------------------------------------------------
 -- Monad/Applicative instances
 
-instance Monad (GenHaxl u) where
+instance Monad (GenHaxl u w) where
   return a = GenHaxl $ \_env -> return (Done a)
   GenHaxl m >>= k = GenHaxl $ \env -> do
     e <- m env
@@ -557,7 +636,7 @@ instance Monad (GenHaxl u) where
   -- We really want the Applicative version of >>
   (>>) = (*>)
 
-instance Functor (GenHaxl u) where
+instance Functor (GenHaxl u w) where
   fmap f (GenHaxl m) = GenHaxl $ \env -> do
     r <- m env
     case r of
@@ -566,7 +645,7 @@ instance Functor (GenHaxl u) where
       Blocked ivar cont -> trace_ "fmap Blocked" $
         return (Blocked ivar (f :<$> cont))
 
-instance Applicative (GenHaxl u) where
+instance Applicative (GenHaxl u w) where
   pure = return
   GenHaxl ff <*> GenHaxl aa = GenHaxl $ \env -> do
     rf <- ff env
@@ -628,12 +707,12 @@ instance Applicative (GenHaxl u) where
 -- Env utils
 
 -- | Extracts data from the 'Env'.
-env :: (Env u -> a) -> GenHaxl u a
+env :: (Env u w -> a) -> GenHaxl u w a
 env f = GenHaxl $ \env -> return (Done (f env))
 
 -- | Returns a version of the Haxl computation which always uses the
 -- provided 'Env', ignoring the one specified by 'runHaxl'.
-withEnv :: Env u -> GenHaxl u a -> GenHaxl u a
+withEnv :: Env u w -> GenHaxl u w a -> GenHaxl u w a
 withEnv newEnv (GenHaxl m) = GenHaxl $ \_env -> do
   r <- m newEnv
   case r of
@@ -647,10 +726,10 @@ withEnv newEnv (GenHaxl m) = GenHaxl $ \_env -> do
 -- Exceptions
 
 -- | Throw an exception in the Haxl monad
-throw :: (Exception e) => e -> GenHaxl u a
+throw :: (Exception e) => e -> GenHaxl u w a
 throw e = GenHaxl $ \_env -> raise e
 
-raise :: (Exception e) => e -> IO (Result u a)
+raise :: (Exception e) => e -> IO (Result u w a)
 raise e
 #ifdef PROFILING
   | Just (HaxlException Nothing h) <- fromException somex = do
@@ -663,7 +742,7 @@ raise e
     somex = toException e
 
 -- | Catch an exception in the Haxl monad
-catch :: Exception e => GenHaxl u a -> (e -> GenHaxl u a) -> GenHaxl u a
+catch :: Exception e => GenHaxl u w a -> (e -> GenHaxl u w a) -> GenHaxl u w a
 catch (GenHaxl m) h = GenHaxl $ \env -> do
    r <- m env
    case r of
@@ -674,20 +753,20 @@ catch (GenHaxl m) h = GenHaxl $ \env -> do
 
 -- | Catch exceptions that satisfy a predicate
 catchIf
-  :: Exception e => (e -> Bool) -> GenHaxl u a -> (e -> GenHaxl u a)
-  -> GenHaxl u a
+  :: Exception e => (e -> Bool) -> GenHaxl u w a -> (e -> GenHaxl u w a)
+  -> GenHaxl u w a
 catchIf cond haxl handler =
   catch haxl $ \e -> if cond e then handler e else throw e
 
 -- | Returns @'Left' e@ if the computation throws an exception @e@, or
 -- @'Right' a@ if it returns a result @a@.
-try :: Exception e => GenHaxl u a -> GenHaxl u (Either e a)
+try :: Exception e => GenHaxl u w a -> GenHaxl u w (Either e a)
 try haxl = (Right <$> haxl) `catch` (return . Left)
 
 -- | @since 0.3.1.0
-instance Catch.MonadThrow (GenHaxl u) where throwM = Haxl.Core.Monad.throw
+instance Catch.MonadThrow (GenHaxl u w) where throwM = Haxl.Core.Monad.throw
 -- | @since 0.3.1.0
-instance Catch.MonadCatch (GenHaxl u) where catch = Haxl.Core.Monad.catch
+instance Catch.MonadCatch (GenHaxl u w) where catch = Haxl.Core.Monad.catch
 
 
 -- -----------------------------------------------------------------------------
@@ -695,14 +774,14 @@ instance Catch.MonadCatch (GenHaxl u) where catch = Haxl.Core.Monad.catch
 
 -- | Under ordinary circumstances this is unnecessary; users of the Haxl
 -- monad should generally /not/ perform arbitrary IO.
-unsafeLiftIO :: IO a -> GenHaxl u a
+unsafeLiftIO :: IO a -> GenHaxl u w a
 unsafeLiftIO m = GenHaxl $ \_env -> Done <$> m
 
 -- | Convert exceptions in the underlying IO monad to exceptions in
 -- the Haxl monad.  This is morally unsafe, because you could then
 -- catch those exceptions in Haxl and observe the underlying execution
 -- order.  Not to be exposed to user code.
-unsafeToHaxlException :: GenHaxl u a -> GenHaxl u a
+unsafeToHaxlException :: GenHaxl u w a -> GenHaxl u w a
 unsafeToHaxlException (GenHaxl m) = GenHaxl $ \env -> do
   r <- m env `Exception.catch` \e -> return (Throw e)
   case r of
@@ -714,7 +793,7 @@ unsafeToHaxlException (GenHaxl m) = GenHaxl $ \env -> do
 -- hierarchy.  Uses 'unsafeToHaxlException' internally.  Typically
 -- this is used at the top level of a Haxl computation, to ensure that
 -- all exceptions are caught.
-tryToHaxlException :: GenHaxl u a -> GenHaxl u (Either HaxlException a)
+tryToHaxlException :: GenHaxl u w a -> GenHaxl u w (Either HaxlException a)
 tryToHaxlException h = left asHaxlException <$> try (unsafeToHaxlException h)
 
 
@@ -724,19 +803,20 @@ tryToHaxlException h = left asHaxlException <$> try (unsafeToHaxlException h)
 -- compiled and run, will recreate the same cache contents.  For
 -- example, the generated code looks something like this:
 --
--- > loadCache :: GenHaxl u ()
+-- > loadCache :: GenHaxl u w ()
 -- > loadCache = do
 -- >   cacheRequest (ListWombats 3) (Right ([1,2,3]))
 -- >   cacheRequest (CountAardvarks "abcabc") (Right (2))
 --
-dumpCacheAsHaskell :: GenHaxl u String
-dumpCacheAsHaskell = dumpCacheAsHaskellFn "loadCache" "GenHaxl u ()"
+dumpCacheAsHaskell :: GenHaxl u w String
+dumpCacheAsHaskell = dumpCacheAsHaskellFn "loadCache" "GenHaxl u w ()"
 
 -- | Dump the contents of the cache as Haskell code that, when
 -- compiled and run, will recreate the same cache contents.
+-- Does not take into account the writes done as part of the computation.
 --
 -- Takes the name and type for the resulting function as arguments.
-dumpCacheAsHaskellFn :: String -> String -> GenHaxl u String
+dumpCacheAsHaskellFn :: String -> String -> GenHaxl u w String
 dumpCacheAsHaskellFn fnName fnType = do
   ref <- env cacheRef  -- NB. cacheRef, not memoRef.  We ignore memoized
                        -- results when dumping the cache.
@@ -744,8 +824,8 @@ dumpCacheAsHaskellFn fnName fnType = do
     readIVar (IVar ref) = do
       r <- readIORef ref
       case r of
-        IVarFull (Ok a) -> return (Just (Right a))
-        IVarFull (ThrowHaxl e) -> return (Just (Left e))
+        IVarFull (Ok a _) -> return (Just (Right a))
+        IVarFull (ThrowHaxl e _) -> return (Just (Left e))
         IVarFull (ThrowIO e) -> return (Just (Left e))
         IVarEmpty _ -> return Nothing