+++ Haxl/DataSource/ConcurrentIO.hs
@@ -21,7 +21,7 @@
 --
 -- For example, to make a concurrent sleep operation:
 --
--- > sleep :: Int -> GenHaxl u Int
+-- > sleep :: Int -> GenHaxl u w Int
 -- > sleep n = dataFetch (Sleep n)
 -- >
 -- > data Sleep